[gcc] CinematicEffects: Remove useless casts

+++ src/cinematic/CinematicEffects.cpp
@@ -203,13 +203,13 @@ void FX_DreamPrecalc(CinematicBitmap * bi, float amp, float fps) {
 			a -= 15.f;
 			a2 += 8.f;
 			
-			nn.x = ((float)n.x) + s.x;
-			nn.y = ((float)n.y) + s.y;
+			nn.x = float(n.x) + s.x;
+			nn.y = float(n.y) + s.y;
 			
-			*t++ = (float)(-o.x + amp * ((2 * (std::sin(nn.x / 20) + std::sin(nn.x * nn.y / 2000)
+			*t++ = (-o.x + amp * ((2 * (std::sin(nn.x / 20) + std::sin(nn.x * nn.y / 2000)
 			                                  + std::sin((nn.x + nn.y) / 100) + std::sin((nn.y - nn.x) / 70) + std::sin((nn.x + 4 * nn.y) / 70)
 			                                  + 2 * std::sin(hypotf(256 - nn.x, (150 - nn.y / 8)) / 40)))));
-			*t++ = (float)(-o.y + amp * (((std::cos(nn.x / 31) + std::cos(nn.x * nn.y / 1783) +
+			*t++ = (-o.y + amp * (((std::cos(nn.x / 31) + std::cos(nn.x * nn.y / 1783) +
 			                              + 2 * std::cos((nn.x + nn.y) / 137) + std::cos((nn.y - nn.x) / 55) + 2 * std::cos((nn.x + 8 * nn.y) / 57)
 			                              + std::cos(hypotf(384 - nn.x, (274 - nn.y / 9)) / 51)))));
